+++ nr/Cargo.toml
@@ -41,4 +41,6 @@ env_logger = "0.9.0"
 rand = {version = "0.8", features = ["small_rng"]}
 
 [features]
+default = ["unstable", "async"]
 unstable = []
+async = []
